Adding comment system and more.

+++ www/block.php
@@ -4,12 +4,13 @@ require_once(__DIR__ . '/../include/head.php');
 $blockId = isset($_GET["blockId"]) ? $_GET["blockId"] : -1;
 
 if(isset($_GET['op']) && $_GET['op'] == 'delete' && isset($_GET['id'])) {
-    $stmt = $mysqli->prepare('DELETE FROM block WHERE id = ?');
-    $stmt->bind_param("i", $_GET['id']);
+    $stmt = $mysqli->prepare('DELETE FROM block WHERE id = ? AND userId = ?');
+    $stmt->bind_param("ii", $_GET['id'], $user['id']);
     $stmt->execute();
 }
 
-$stmt = $mysqli->prepare('SELECT * FROM category');
+$stmt = $mysqli->prepare('SELECT * FROM category WHERE userId = ?');
+$stmt->bind_param("i", $user['id']);
 $stmt->execute();
 $categories = fetchAssocAll($stmt, 'id');
 
@@ -65,8 +66,9 @@ $types = [
         <div class="container">
             <h3>Block editor</h3>
             <div class="row">
-                <a class="button" href="listVideos.php">Home</a>
+                <a class="button" href="listVideos.php">List videos</a>
                 <a class="button" href="category.php">Categories</a>
+                <a class="button" href="comments.php">List comments</a>
             </div>
 
             <div class="row">
